refactor: image, restaurants를 constants 폴더 밑으로 이동
