+++ src/constants/image.ts
@@ -1,0 +1,15 @@
+import asian from '../../assets/category-asian.png';
+import chinese from '../../assets/category-chinese.png';
+import etc from '../../assets/category-etc.png';
+import japanese from '../../assets/category-japanese.png';
+import korean from '../../assets/category-korean.png';
+import western from '../../assets/category-western.png';
+
+export const imgSrc = {
+  한식: korean,
+  중식: chinese,
+  일식: japanese,
+  양식: western,
+  아시안: asian,
+  기타: etc,
+};
